Fixed marching on the ATV (really odd that you can tell the march to display selection 1.4 and it does!).
git-svn-id: https://appletv.nanopi.net/svn/trunk@280 7e1782f7-7b3f-45d9-b215-e5100ed717cc

+++ SapphirePosterChooser.m
@@ -391,7 +391,7 @@ NSData *CreateBitmapDataFromImage(CGImageRef image, unsigned int width, unsigned
 	[SapphireFrontRowCompat addSublayer:posterMarch toControl:self];
 }
 
-- (void)setSelectionForPoster:(int)sel
+- (void)setSelectionForPoster:(double)sel
 {
 	NSMethodSignature *signature = [posterMarch methodSignatureForSelector:@selector(setSelection:)];
 	NSInvocation *selInv = [NSInvocation invocationWithMethodSignature:signature];
